Correct binary conversion

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -9,7 +9,7 @@
 unsigned int binary_to_uint(const char *b)
 {
 	int i, pow = 1;
-	unsigned int res, ret, remd;
+	unsigned long int res, ret, remd;
 
 	ret = 0;
 	if (b == NULL)
